Allow generic op implementations (#14)
This adds the ability for op implementations to be generic by expanding the macros to handle generic parameters just before the lambda. This implementation prevents duplication of the rules and allows for arbitrarily complex generic parameters and constraints (including lifetimes).

It does have the potential downside of increasing the macro expansion recursion depth. In the existing non-generic case, it should just add a constant 2 calls to the max depth. When using generic parameters it adds `2 + N` where `N` is the number of tokens in the generic parameters, which could be decently high. This is due to the ["tt muncher"](https://veykril.github.io/tlborm/decl-macros/patterns/tt-muncher.html) strategy taken; see the docs on `_generic_params_shifter_internal` for more details. However in the case that consumers of the crate run into issues with maximum recursion limit here, they can opt into a higher limit (I would definitely consider it an edge case).

This change should be fully backwards compatible. As far as I can tell, all existing use cases should be unchanged except for the 2 additional calls in the macro stack. Given that the macro recursion limit defaults to 128, I'd be very surprised if the increase causes things to break for folks.

The change also includes tests for the new functionality and updates to the documentation.

Inspired by https://github.com/carbotaniuman/auto_ops/pull/4.